Refactor to use Utils
