+++ lib/main.js
@@ -29,11 +29,11 @@ var _coPrompt = require('co-prompt');
 
 var _coPrompt2 = _interopRequireDefault(_coPrompt);
 
-var _default = require('../templates/default.json');
+var _utils = require('./utils');
 
-var _default2 = _interopRequireDefault(_default);
+var Utils = _interopRequireWildcard(_utils);
 
-var _child_process = require('child_process');
+function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) newObj[key] = obj[key]; } } newObj.default = obj; return newObj; } }
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -41,230 +41,20 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 // Libs
 
 
-var emptyFunc = () => {};
-
-function makeNote(jsonObj) {
-  let cb = arguments.length <= 1 || arguments[1] === undefined ? emptyFunc : arguments[1];
-
-  var dir = getDateDir();
-  var toMd = `${ dir }/note.md`;
-  var noteMd = _fsExtra2.default.createWriteStream(toMd);
-
-  Object.keys(jsonObj).map(function (title) {
-    var cliRef = jsonObj[title]['cli-ref'];
-    noteMd.write(`# ${ title } --> ${ cliRef }\n`);
-    Object.keys(jsonObj[title]['items']).map(function (items, index) {
-      var status = jsonObj[title]['items'][items]['status'];
-      var checkBox = '- [ ]';
-      var itemIndex = ` ${ index }.) `;
-      if (status === 'complete') {
-        checkBox = '- [x]';
-      } else if (status === 'failed') {
-        checkBox = '- [-]';
-      }
-      noteMd.write(checkBox + itemIndex + jsonObj[title]['items'][items]['description'] + "\n");
-    });
-    noteMd.write("\n");
-  });
-  noteMd.on('finish', cb);
-  noteMd.end();
-}
-
-function removeSection(section) {
-  var dir = getDateDir();
-  var toData = `${ dir }/data.json`;
-  var dataJSON = _fsExtra2.default.readJsonSync(toData);
-
-  Object.keys(dataJSON).map(note => {
-    if (dataJSON[note]['cli-ref'] === section) {
-      console.log(`"${ _chalk2.default.cyan(note) }" section was removed!`);
-      delete dataJSON[note];
-    }
-  });
-
-  _fsExtra2.default.writeJsonSync(toData, dataJSON);
-  makeNote(dataJSON);
-}
-
-function createSection(name, cliRef) {
-  let description = arguments.length <= 2 || arguments[2] === undefined ? 'no description' : arguments[2];
-
-  var dir = getDateDir();
-  var toData = `${ dir }/data.json`;
-  var dataJSON = _fsExtra2.default.readJsonSync(toData);
-  dataJSON[name] = {
-    'cli-ref': cliRef,
-    'description': description,
-    'items': []
-  };
-  _fsExtra2.default.writeJsonSync(toData, dataJSON);
-  makeNote(dataJSON, () => {
-    process.exit();
-  });
-}
-
-function addNote(noteObj, key) {
-  const today = (0, _moment2.default)().format("DD-MM-YYYY");
-  var configPath = `${ process.env.HOME }/.nonoterc.json`;
-  _fsExtra2.default.readJson(configPath, (err, configJSON) => {
-    if (err) console.error(err);
-
-    const notesDir = configJSON.notesDirectory;
-    const days = `${ notesDir }/days/`;
-    const toDir = days + today;
-    const dataPath = `${ toDir }/data.json`;
-
-    var noteString = noteObj.reduce(function (memo, word) {
-      return memo + ' ' + word;
-    });
-
-    var descriptionObj = {
-      description: noteString,
-      status: 'incomplete'
-    };
-
-    // modify toData
-    _fsExtra2.default.readJson(dataPath, (err, dataJSON) => {
-      Object.keys(dataJSON).map(function (note) {
-        if (dataJSON[note]['cli-ref'] === key) {
-          dataJSON[note]['items'].push(descriptionObj);
-          _fsExtra2.default.writeJson(dataPath, dataJSON, err => {
-            if (err) return console.error(err);
-
-            makeNote(dataJSON, () => {
-              process.exit();
-            });
-          });
-        }
-      });
-    });
-  });
-}
-
-function changeStatus(index, key, cb) {
-  const today = (0, _moment2.default)().format("DD-MM-YYYY");
-  var configPath = `${ process.env.HOME }/.nonoterc.json`;
-  _fsExtra2.default.readJson(configPath, (err, configJSON) => {
-    if (err) console.error(err);
-
-    const notesDir = configJSON.notesDirectory;
-    const days = `${ notesDir }/days/`;
-    const toDir = days + today;
-    var toData = `${ toDir }/data.json`;
-
-    _fsExtra2.default.readJson(toData, (err, dataJSON) => {
-      if (err) console.error(err);
-
-      var cliFound = false;
-      Object.keys(dataJSON).map(function (note, noteIndex) {
-        if (dataJSON[note]['cli-ref'] === key) {
-          cliFound = true;
-          if (!dataJSON[note]['items'][index]) {
-            throw new Error(`index ${ index } in "${ key }" object does not exist`);
-          }
-          cb(dataJSON[note]['items'], index);
-          _fsExtra2.default.writeJson(toData, dataJSON, err => {
-            if (err) return console.error(err);
-
-            makeNote(dataJSON, () => {
-              process.exit();
-            });
-          });
-        } else if (Object.keys(dataJSON).length === noteIndex + 1 && !cliFound) {
-          throw new Error(`"${ key }" <cli-ref> does not exist`);
-        }
-      });
-    });
-  });
-}
-
-function removeNote(arry, index) {
-  arry.splice(index, 1);
-}
-
-function completeNote(arry, index) {
-  arry[index]['status'] = 'complete';
-}
-
-function incompleteNote(arry, index) {
-  arry[index]['status'] = 'incomplete';
-}
-
-function failNote(arry, index) {
-  arry[index]['status'] = 'failed';
-}
-
-function initializeNotes(userDir) {
-  var lastChar = userDir.substr(userDir.length - 1);
-  if (lastChar === "/") {
-    userDir = userDir.slice(0, -1);
-  }
-
-  var doubleIndex = userDir.indexOf("//");
-  if (doubleIndex > -1) {
-    userDir = userDir.slice(0, doubleIndex) + userDir.slice(doubleIndex + 1);
-  }
-
-  var rcFile = `${ process.env.HOME }/.nonoterc.json`;
-  _fsExtra2.default.closeSync(_fsExtra2.default.openSync(rcFile, 'w'));
-
-  var dotFileJSON = {};
-  dotFileJSON.notesDirectory = userDir;
-
-  _fsExtra2.default.writeJsonSync(rcFile, dotFileJSON);
-  console.log(_chalk2.default.green('Success!'));
-  console.log(' ');
-  console.log('dotfile `.nonoterc.json` created at $HOME');
-  console.log(' ');
-  console.log('Notes will be made in this directory: ');
-  console.log(_chalk2.default.cyan(dotFileJSON.notesDirectory));
-  console.log(' ');
-}
-
-function getRootDir() {
-  var config = `${ process.env.HOME }/.nonoterc.json`;
-  // TODO: handle the case of `nonote new` when `nonote init` has not been run
-  return _fsExtra2.default.readJsonSync(config).notesDirectory;
-}
-
-function getDateDir() {
-  const today = (0, _moment2.default)().format("DD-MM-YYYY");
-  const notesDir = getRootDir();
-  const days = `${ notesDir }/days/`;
-  const toDir = days + today;
-  return toDir;
-}
-
-function createDir(create, path) {
-  var templateDest = path + '/templates/';
-  var templateFile = path + '/templates/default.json';
-  if (create) {
-    _fsExtra2.default.mkdirsSync(path);
-    _fsExtra2.default.mkdirsSync(templateDest);
-    _fsExtra2.default.closeSync(_fsExtra2.default.openSync(templateFile, 'w'));
-    _fsExtra2.default.writeJsonSync(templateFile, _default2.default);
-    console.log(' ');
-    console.log(_chalk2.default.green('Success!'));
-  } else {
-    console.log(' ');
-    console.log('Make sure that dir exists and has a templates dir in it with a note config!');
-  }
-}
-
 _commander2.default.version('0.0.1').command('new [template]').alias('n').description('create a new note for the day').action(function (template, cmd) {
   if (!template) {
     template = 'default';
   }
-  const notePath = getDateDir();
+  const notePath = Utils.getDateDir();
   const noteJSON = `${ notePath }/data.json`;
-  const rootDir = getRootDir();
+  const rootDir = Utils.getRootDir();
   const templateData = `${ rootDir }/templates/${ template }.json`;
 
   console.log(`creating new note from the ${ _chalk2.default.cyan(template) } template!`);
 
   _fsExtra2.default.mkdirsSync(notePath);
   _fsExtra2.default.copySync(templateData, noteJSON);
-  makeNote(_fsExtra2.default.readJsonSync(noteJSON));
+  Utils.makeNote(_fsExtra2.default.readJsonSync(noteJSON));
 
   console.log(`\nnew note created for today: \n${ _chalk2.default.cyan(notePath) }`);
 });
@@ -302,7 +92,7 @@ _commander2.default.command('section').alias('s').description('create new sectio
     const cliRefName = yield (0, _coPrompt2.default)('Cli reference key: ');
     const description = yield (0, _coPrompt2.default)('Short description: ');
 
-    createSection(sectionName, cliRefName, description);
+    Utils.createSection(sectionName, cliRefName, description);
 
     console.log(_chalk2.default.green(`new section '${ sectionName }' was created!`));
   });
@@ -310,13 +100,12 @@ _commander2.default.command('section').alias('s').description('create new sectio
 
 _commander2.default.command('init').description('initializes notes').action(function () {
   (0, _co2.default)(function* () {
-
     const notesDirPath = yield (0, _coPrompt2.default)('notes directory path (from $HOME): ');
     const homePath = `${ process.env.HOME }/${ notesDirPath }`;
-    initializeNotes(homePath);
+    Utils.initializeNotes(homePath);
 
     const shouldCreate = yield _coPrompt2.default.confirm(`Would you like me to create "${ homePath }" for you?(Recommended) [y/N] `);
-    createDir(shouldCreate, homePath);
+    Utils.createDir(shouldCreate, homePath);
     console.log('Start taking notes with', _chalk2.default.cyan('nonote new!'));
     // TODO: create a readme
     process.exit();
@@ -325,7 +114,7 @@ _commander2.default.command('init').description('initializes notes').action(func
 
 _commander2.default.command('add <cli-ref> [notes...]').alias('a').description('add note to object').action(function (ref, note) {
   try {
-    addNote(note, ref);
+    Utils.addNote(note, ref);
     console.log(_chalk2.default.green('note added!'));
   } catch (e) {
     console.log(_chalk2.default.red(e));
@@ -334,10 +123,10 @@ _commander2.default.command('add <cli-ref> [notes...]').alias('a').description('
 
 _commander2.default.command('remove <cli-ref> [index]').alias('r').option('-s, --section', 'remove a seciton').description('remove note from note object').action(function (ref, note, options) {
   if (options.section && note === undefined) {
-    removeSection(ref);
+    Utils.removeSection(ref);
   } else {
     try {
-      changeStatus(note, ref, removeNote);
+      Utils.changeStatus(note, ref, Utils.removeNote);
       console.log(_chalk2.default.green(`note at index[${ note }] was removed!`));
     } catch (e) {
       console.log(_chalk2.default.red(e));
@@ -347,7 +136,7 @@ _commander2.default.command('remove <cli-ref> [index]').alias('r').option('-s, -
 
 _commander2.default.command('complete <cli-ref> [index]').alias('c').description('mark item as complete').action(function (ref, note) {
   try {
-    changeStatus(note, ref, completeNote);
+    Utils.changeStatus(note, ref, Utils.completeNote);
     console.log(_chalk2.default.green(`note at index[${ note }] was marked as complete!`));
   } catch (e) {
     console.log(_chalk2.default.red(e));
@@ -356,7 +145,7 @@ _commander2.default.command('complete <cli-ref> [index]').alias('c').description
 
 _commander2.default.command('incomplete <cli-ref> [index]').alias('i').description('mark item as incomplete').action(function (ref, note) {
   try {
-    changeStatus(note, ref, incompleteNote);
+    Utils.changeStatus(note, ref, Utils.incompleteNote);
     console.log(_chalk2.default.green(`note at index[${ note }] was marked as incomplete!`));
   } catch (e) {
     console.log(_chalk2.default.red(e));
@@ -365,7 +154,7 @@ _commander2.default.command('incomplete <cli-ref> [index]').alias('i').descripti
 
 _commander2.default.command('failed [cli-ref] <index>').alias('f').description('mark item as failed').action(function (ref, note) {
   try {
-    changeStatus(note, ref, failNote);
+    Utils.changeStatus(note, ref, Utils.failNote);
     console.log(_chalk2.default.green(`note at index[${ note }] was marked as failed :(`));
   } catch (e) {
     console.log(_chalk2.default.red(e));
