+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nonote",
-  "version": "1.2.7",
+  "version": "1.3.0",
   "description": "a node application for taking notes in markdown from the command line",
   "main": "./lib/main.js",
   "scripts": {
@@ -40,7 +40,6 @@
     "commander": "^2.9.0",
     "fs-extra": "^0.26.7",
     "jsonfile": "^2.2.3",
-    "mkdirp": "^0.5.1",
     "moment": "^2.12.0",
     "source-map-support": "^0.4.0"
   },
